bots/learning_bot/requirements.txt
python-dotenv==1.0.0
discord.py==2.3.1
langchain==0.1.0
faiss-cpu==1.7.4
chromadb==0.4.22
pypdf==3.17.1
python-docx==1.0.1
openrouter==1.0
pytz==2023.3
python-dateutil==2.8.2
aiohttp==3.9.1
apscheduler==3.10.4 